move builder and handler into same file
